+++ .editorconfig
@@ -1,0 +1,33 @@
+﻿root = true
+
+[*]
+charset = utf-8-bom
+end_of_line = crlf
+indent_size = 4
+indent_style = space
+insert_final_newline = true
+tab_width = 4
+trim_trailing_whitespace = true
+
+# Xml project files
+[*.*proj]
+indent_size = 2
+
+# Xml build files
+[*.builds]
+indent_size = 2
+
+# Xml files
+[*.{xml,stylecop,resx,ruleset,.xsd}]
+indent_size = 2
+
+# Xml config files
+[*.{props,targets,config,nuspec}]
+indent_size = 2
+
+[*.{js,json}]
+indent_size = 2
+
+# Shell scripts
+[*.sh]
+end_of_line = lf
