--- global.json
@@ -1,5 +1,0 @@
-{
-  "sdk": {
-    "version": "1.0.0-preview2-003131"
-  }
-}
