+++ src/StatsdClient/StatsdClient.csproj
@@ -1,0 +1,29 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <Description>Statsd client for C#, providing a full set of counter/timer/gauge/set functionality in an easy to use static. Helpers are provided for easily timing chunks of code too.</Description>
+    <AssemblyDescription>Statsd Client</AssemblyDescription>
+    <AssemblyTitle>Statsd C# Client</AssemblyTitle>
+    <Authors>Goncalo Pereira;Darrell Mozingo</Authors>
+    <TargetFrameworks>net45;net471;netstandard1.3;netstandard2.0</TargetFrameworks>
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+    <AssemblyName>StatsdClient</AssemblyName>
+    <PackageId>StatsdClient.Async</PackageId>
+    <PackageTags>stats;statsd;metrics</PackageTags>
+    <PackageReleaseNotes>See CHANGELOG.md in the project's GitHub repository.</PackageReleaseNotes>
+    <PackageProjectUrl>https://github.com/pengweiqhca/statsd-csharp-client</PackageProjectUrl>
+    <PackageLicenseUrl>https://github.com/pengweiqhca/statsd-csharp-client/blob/master/MIT-LICENSE.md</PackageLicenseUrl>
+    <RepositoryUrl>https://github.com/pengweiqhca/statsd-csharp-client</RepositoryUrl>
+    <SignAssembly>true</SignAssembly>
+    <AssemblyOriginatorKeyFile>..\..\key.snk</AssemblyOriginatorKeyFile>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="ConfigureAwaitChecker.Analyzer" Version="2.*" Condition="'$(Configuration)' != 'Release'" PrivateAssets="All" />
+  </ItemGroup>
+
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard1.3'">
+    <PackageReference Include="System.Net.NameResolution" Version="4.0.0" />
+  </ItemGroup>
+
+</Project>
